Datetime output format, query and object patch
