+++ init.rb
@@ -3,14 +3,16 @@ Redmine::Plugin.register :computed_custom_field do
   author 'Yakov Annikov'
   url 'https://github.com/annikoff/redmine_plugin_computed_custom_field'
   description ''
-  version '0.0.2'
+  version '0.0.3'
   settings :default => {}
 end
 
 ActionDispatch::Callbacks.to_prepare do
   require 'patcher'
-  require 'patches/custom_field_patch'
-  require 'patches/field_format_patch'
+  require 'patches/custom_field'
+  require 'patches/field_format'
+  require 'patches/query'
+  require 'patches/object'
   require 'patches/klass_patch'
 end
